+++ src/utils/response/log_functions_http_parsing/log_functions_http_parsing.h
@@ -8,10 +8,10 @@
 #include "utils/parsers/http/parser_request.h"
 #include "utils/response/response.h"
 
-void log_request(struct vhost *vhost, struct request *req, size_t *status_code,
+void log_request(struct vhost *vhost, struct request *req, int *status_code,
                  ssize_t index);
 
-void log_response(struct vhost *vhost, struct request *req, size_t *status_code,
+void log_response(struct vhost *vhost, struct request *req, int *status_code,
                   ssize_t index);
 
 #endif /* !LOG_FUNCTIONS_HTTP_PARSING_H */
